stage3データ追加 マップ上に砲台を表示できる処理追加

+++ ACG/ACG/ObjCannon.h
@@ -11,7 +11,7 @@ using namespace GameL;
 class CObjCannon :public CObj
 {
 public:
-	CObjCannon(float x,float y); 	//コンストラクタ
+	CObjCannon(int x,int y); 	//コンストラクタ
 	~CObjCannon() {};			//デストラクタ
 	void Init();				//イニシャライズ
 	void Action();				//アクション
